Models are the same for embedded and desktop - no need to split in folders
