+++ .gitignore
@@ -7,7 +7,7 @@ cmake-build*
 logs/
 results/
 venv/
-extras/desktop/models/stateful-rnn/stateful-lstm/
-extras/desktop/models/hybrid-nn/GuitarLSTM/
-extras/desktop/models/cnn/steerable-nafx/
-extras/desktop/models/model-pool/example-models/
+extras/models/stateful-rnn/stateful-lstm/
+extras/models/hybrid-nn/GuitarLSTM/
+extras/models/cnn/steerable-nafx/
+extras/models/model-pool/example-models/
