+++ examples/desktop/minimal-inference/onnxruntime/minimal-onnxruntime.cpp
@@ -8,11 +8,11 @@ Licence: MIT
 #include <iostream>
 #include <onnxruntime_cxx_api.h>
 
-#include "../../../../extras/desktop/models/stateful-rnn/StatefulRNNConfig.h"
-#include "../../../../extras/desktop/models/hybrid-nn/HybridNNConfig.h"
-#include "../../../../extras/desktop/models/cnn/CNNConfig.h"
-#include "../../../../extras/desktop/models/model-pool/SimpleGainConfig.h"
-#include "../../../../extras/desktop/models/model-pool/SimpleStereoGainConfig.h"
+#include "../../../../extras/models/stateful-rnn/StatefulRNNConfig.h"
+#include "../../../../extras/models/hybrid-nn/HybridNNConfig.h"
+#include "../../../../extras/models/cnn/CNNConfig.h"
+#include "../../../../extras/models/model-pool/SimpleGainConfig.h"
+#include "../../../../extras/models/model-pool/SimpleStereoGainConfig.h"
 
 #include "../../../../include/anira/utils/MemoryBlock.h"
 #include "../../../../include/anira/utils/AudioBuffer.h"
