+++ extras/models/stateful-rnn/StatefulRNNConfig.h
